chore: mydashboard 타입과 api 로직 분리
